Fixes bugs around getting the version.
Version_Get in lib-ci would always create a new version
number. Unfortunately, in some builds it is used multiple times causing
several different version numbers to be used.

This moves the version creation part in the setup-version.sh script,
which supports for `sbt` and `txt` mode. It creates a CI version file
that is then used by `Version_Get` to always return the same version
number.

This closes #35.